Config gulp and es-lint
