+++ package.json
@@ -1,0 +1,33 @@
+{
+  "name": "magento1_plugin",
+  "version": "1.0.0",
+  "description": "Ebanx magento1 plugin",
+  "repository": "git@github.com:ebanx/magento-gateway-ebanx.git",
+  "author": "rafabruno",
+  "main": "gulpfile.js",
+  "scripts": {
+    "start": "npm run build",
+    "build": "gulp default",
+    "watch": "gulp watch",
+    "lint": "gulp lint"
+  },
+  "devDependencies": {
+    "babel-preset-env": "^1.6.1",
+    "babelify": "^7.3.0",
+    "eslint-restricted-globals": "^0.1.1",
+    "gulp": "^3.9.1",
+    "gulp-autoprefixer": "^4.0.0",
+    "gulp-babel": "^7.0.0",
+    "gulp-browserify": "^0.5.1",
+    "gulp-clean-css": "^3.9.0",
+    "gulp-eslint": "^4.0.1",
+    "gulp-jshint": "^2.0.4",
+    "gulp-rename": "^1.2.2",
+    "gulp-sass": "^3.1.0",
+    "gulp-sourcemaps": "^2.6.1",
+    "gulp-uglify": "^3.0.0",
+    "install": "^0.10.1",
+    "jshint": "^2.9.5",
+    "vanilla-masker": "^1.2.0"
+  }
+}
