+++ skin/frontend/base/default/ebanx/js/amscheckout-inputs.js
@@ -1,3 +1,2 @@
-var amsCheckoutHandler = function () {
-  return qs('label[for="billing\\:taxvat_number"]');
-};
+"use strict";!function r(e,n,t){function o(i,f){if(!n[i]){if(!e[i]){var c="function"==typeof require&&require;if(!f&&c)return c(i,!0);if(u)return u(i,!0);throw new Error("Cannot find module '"+i+"'")}var a=n[i]={exports:{}};e[i][0].call(a.exports,function(r){var n=e[i][1][r];return o(n||r)},a,a.exports,r,e,n,t)}return n[i].exports}for(var u="function"==typeof require&&require,i=0;i<t.length;i++)o(t[i]);return o}({1:[function(r,e,n){},{}]},{},[1]);
+//# sourceMappingURL=amscheckout-inputs.js.map
