+++ skin/frontend/base/default/ebanx/js/checkout.js
@@ -1,94 +1,2 @@
-function handleEbanxForm(countryCode, paymentType) {
-  var getById = function (element) {
-    return document.getElementById(element);
-  };
-  var responseData = null;
-
-  var cardName = getById('ebanx_' + paymentType + '_' + countryCode + '_' + paymentType + '_name');
-  var cardNumber = getById('ebanx_' + paymentType + '_' + countryCode + '_' + paymentType + '_number');
-  var cardExpirationMonth = getById('ebanx_' + paymentType + '_' + countryCode + '_expiration');
-  var cardExpirationYear = getById('ebanx_' + paymentType + '_' + countryCode + '_expiration_yr');
-  var cardCvv = getById('ebanx_' + paymentType + '_' + countryCode + '_' + paymentType + '_cid');
-  var ebanxToken = getById('ebanx_' + paymentType + '_' + countryCode + '_token');
-  var ebanxBrand = getById('ebanx_' + paymentType + '_' + countryCode + '_brand');
-  var ebanxMaskedCardNumber = getById('ebanx_' + paymentType + '_' + countryCode + '_masked_card_number');
-  var ebanxDeviceFingerprint = getById('ebanx_' + paymentType + '_' + countryCode + '_device_fingerprint');
-  var ebanxMode = getById('ebanx_' + paymentType + '_' + countryCode + '_mode');
-  var ebanxIntegrationKey = getById('ebanx_' + paymentType + '_' + countryCode + '_integration_key');
-  var ebanxCountry = getById('ebanx_' + paymentType + '_' + countryCode + '_country');
-
-  var hasEbanxForm = typeof getById('payment_form_ebanx_' + paymentType + '_' + countryCode) !== 'undefined';
-
-  var mode = ebanxMode.value === 'sandbox' ? 'test' : 'production';
-  EBANX.config.setMode(mode);
-  EBANX.config.setPublishableKey(ebanxIntegrationKey.value);
-  EBANX.config.setCountry(ebanxCountry.value);
-
-  var handleToken = function () {
-    if (!isFormEmpty()) {
-      generateToken();
-    }
-  };
-
-  var isFormEmpty = function () {
-    return !cardNumber.value.length ||
-      !cardName.value.length ||
-      !cardExpirationMonth.value.length ||
-      !cardExpirationYear.value.length ||
-      !cardCvv.value.length;
-  };
-
-  var generateToken = function () {
-    if (!responseData) {
-      var placeOrderButton = document.querySelector('#review-buttons-container > button');
-      if (typeof placeOrderButton !== 'undefined' && placeOrderButton) {
-        placeOrderButton.disabled = true;
-      }
-
-      EBANX.card.createToken({
-        card_number: parseInt(cardNumber.value.replace(/ /g, '')),
-        card_name: cardName.value,
-        card_due_date: (parseInt(cardExpirationMonth.value) || 0) + '/' + (parseInt(cardExpirationYear.value) || 0),
-        card_cvv: cardCvv.value
-      }, saveToken);
-    }
-  };
-
-  var saveToken = function (response) {
-    if (response.data.hasOwnProperty('status')) {
-      responseData = response.data;
-      ebanxToken.value = responseData.token;
-      ebanxBrand.value = responseData.payment_type_code;
-      ebanxMaskedCardNumber.value = responseData.masked_card_number;
-      ebanxDeviceFingerprint.value = responseData.deviceId;
-
-      var placeOrderButton = document.querySelector('#review-buttons-container > button');
-      if (typeof placeOrderButton !== 'undefined' && placeOrderButton) {
-        placeOrderButton.disabled = false;
-      }
-      return;
-    }
-  };
-
-  var clearResponseData = function () {
-    responseData = null;
-    ebanxToken.value = '';
-    ebanxBrand.value = '';
-    ebanxMaskedCardNumber.value = '';
-    ebanxDeviceFingerprint.value = '';
-  };
-
-  if (hasEbanxForm) {
-    cardName.addEventListener('blur', handleToken, false);
-    cardNumber.addEventListener('blur', handleToken, false);
-    cardExpirationMonth.addEventListener('blur', handleToken, false);
-    cardExpirationYear.addEventListener('blur', handleToken, false);
-    cardCvv.addEventListener('blur', handleToken, false);
-
-    cardName.addEventListener('change', clearResponseData, false);
-    cardNumber.addEventListener('change', clearResponseData, false);
-    cardExpirationMonth.addEventListener('change', clearResponseData, false);
-    cardExpirationYear.addEventListener('change', clearResponseData, false);
-    cardCvv.addEventListener('change', clearResponseData, false);
-  }
-}
+"use strict";!function r(e,n,t){function o(i,f){if(!n[i]){if(!e[i]){var c="function"==typeof require&&require;if(!f&&c)return c(i,!0);if(u)return u(i,!0);throw new Error("Cannot find module '"+i+"'")}var a=n[i]={exports:{}};e[i][0].call(a.exports,function(r){var n=e[i][1][r];return o(n||r)},a,a.exports,r,e,n,t)}return n[i].exports}for(var u="function"==typeof require&&require,i=0;i<t.length;i++)o(t[i]);return o}({1:[function(r,e,n){},{}]},{},[1]);
+//# sourceMappingURL=checkout.js.map
