+++ skin/frontend/base/default/ebanx/js/document.js
@@ -1,82 +1,2 @@
-'use strict';
-
-var defaultLabel;
-var taxVatLabel;
-var taxVatInput;
-
-var qs = function (el) {
-  return document.querySelector(el);
-};
-
-function inputHandler(masks, max, event) {
-  var c = event.target;
-  var v = c.value.replace(/\D/g, '');
-  var m = c.value.length > max ? 1 : 0;
-  VMasker(c).unMask();
-  VMasker(c).maskPattern(masks[m]);
-  c.value = VMasker.toPattern(v, masks[m]);
-}
-
-var getLabelByCountry = function (country, defaultLabel) {
-  switch (country.toLowerCase()) {
-    case 'br':
-      return EBANXData.brazilAllowedDocumentFields.join(' / ').toUpperCase();
-    case 'co':
-      return 'DNI';
-    case 'cl':
-      return 'RUT';
-    default:
-      return defaultLabel;
-  }
-};
-
-var changeTaxVatLabel = function () {
-  var country = this.value;
-  var newLabel = getLabelByCountry(country, defaultLabel);
-
-  OSCRequire(country);
-
-  setTimeout(function(){taxVatInput.placeholder = '';}, 10);
-
-  taxVatLabel.innerHTML = newLabel;
-  if(country === 'BR' || country === 'CO' || country === 'CL') {
-      setTimeout(function(){taxVatInput.placeholder = newLabel;}, 10)
-  }
-
-  VMasker(taxVatInput).unMask();
-
-  if (country === 'BR') {
-    var taxVatMask = newLabel.indexOf('CNPJ') !== -1 ? ['999.999.999-99', '99.999.999/9999-99'] : ['999.999.999-99', '999.999.999-99'];
-    VMasker(taxVatInput).maskPattern(taxVatMask[0]);
-    taxVatInput.addEventListener('input', inputHandler.bind(undefined, taxVatMask, 14), false);
-  }
-};
-
-var init = function () {
-  if (!EBANXData.maskTaxVat) {
-    return;
-  }
-
-  var countrySelect = qs('#billing\\:country_id');
-  taxVatLabel = qs('label[for="billing\\:taxvat"]');
-  taxVatInput = document.getElementById('billing:taxvat');
-
-  if (!taxVatLabel && typeof amsCheckoutHandler === 'function') {
-    taxVatLabel = amsCheckoutHandler();
-  }
-
-  if (taxVatLabel && countrySelect && taxVatInput) {
-    defaultLabel = taxVatLabel.innerHTML;
-
-    countrySelect.addEventListener('change', changeTaxVatLabel);
-    countrySelect.dispatchEvent(new Event('change'));
-  }
-};
-
-var OSCRequire = function (country) {
-  if (typeof ebanxRemoveOSCRequireDocument !== 'undefined') {
-    ebanxRemoveOSCRequireDocument(country);
-  }
-};
-
-document.addEventListener('DOMContentLoaded', init);
+"use strict";!function e(n,t,r){function a(i,u){if(!t[i]){if(!n[i]){var c="function"==typeof require&&require;if(!u&&c)return c(i,!0);if(o)return o(i,!0);throw new Error("Cannot find module '"+i+"'")}var l=t[i]={exports:{}};n[i][0].call(l.exports,function(e){var t=n[i][1][e];return a(t||e)},l,l.exports,e,n,t,r)}return t[i].exports}for(var o="function"==typeof require&&require,i=0;i<r.length;i++)a(r[i]);return a}({1:[function(e,n,t){var r,a,o,i=function(e){return document.querySelector(e)};var u=function(){var e=this.value,n=function(e,n){switch(e.toLowerCase()){case"br":return EBANXData.brazilAllowedDocumentFields.join(" / ").toUpperCase();case"co":return"DNI";case"cl":return"RUT";default:return n}}(e,r);if(c(e),setTimeout(function(){o.placeholder=""},10),a.innerHTML=n,"BR"!==e&&"CO"!==e&&"CL"!==e||setTimeout(function(){o.placeholder=n},10),VMasker(o).unMask(),"BR"===e){var t=-1!==n.indexOf("CNPJ")?["999.999.999-99","99.999.999/9999-99"]:["999.999.999-99","999.999.999-99"];VMasker(o).maskPattern(t[0]),o.addEventListener("input",function(e,n,t){var r=t.target,a=r.value.replace(/\D/g,""),o=r.value.length>n?1:0;VMasker(r).unMask(),VMasker(r).maskPattern(e[o]),r.value=VMasker.toPattern(a,e[o])}.bind(void 0,t,14),!1)}},c=function(e){"undefined"!=typeof ebanxRemoveOSCRequireDocument&&ebanxRemoveOSCRequireDocument(e)};document.addEventListener("DOMContentLoaded",function(){if(EBANXData.maskTaxVat){var e=i("#billing\\:country_id");a=i('label[for="billing\\:taxvat"]'),o=document.getElementById("billing:taxvat"),a||"function"!=typeof amsCheckoutHandler||(a=amsCheckoutHandler()),a&&e&&o&&(r=a.innerHTML,e.addEventListener("change",u),e.dispatchEvent(new Event("change")))}})},{}]},{},[1]);
+//# sourceMappingURL=document.js.map
