+++ skin/frontend/base/default/ebanx/js/oneclick.js
@@ -1,33 +1,2 @@
-document.addEventListener('DOMContentLoaded', function () {
-  var button = document.getElementById('product-oneclick-ebanx-button');
-  var tooltip = document.getElementById('ebanx-one-click-tooltip');
-  var close = document.getElementById('ebanx-one-click-close-button');
-  var elements = document.getElementsByName('payment[selected_card]');
-  var payButton = document.getElementById('ebanx-oneclick-pay-button');
-  var form = document.getElementById('product_addtocart_form');
-
-  if (elements.length) {
-    elements[0].checked = true;
-  }
-
-  if (button) {
-    button.addEventListener('click', function (e) {
-      e.preventDefault();
-      tooltip.classList.toggle('is-active');
-    });
-  }
-
-  if (close) {
-    close.addEventListener('click', function (e) {
-      e.preventDefault();
-      tooltip.classList.remove('is-active');
-    });
-  }
-
-  if (payButton && form) {
-    payButton.addEventListener('click', function (e) {
-      form.action = '/ebanx/oneclick/pay';
-      productAddToCartForm.submit(payButton);
-    });
-  }
-});
+"use strict";!function e(t,n,o){function c(i,u){if(!n[i]){if(!t[i]){var d="function"==typeof require&&require;if(!u&&d)return d(i,!0);if(r)return r(i,!0);throw new Error("Cannot find module '"+i+"'")}var a=n[i]={exports:{}};t[i][0].call(a.exports,function(e){var n=t[i][1][e];return c(n||e)},a,a.exports,e,t,n,o)}return n[i].exports}for(var r="function"==typeof require&&require,i=0;i<o.length;i++)c(o[i]);return c}({1:[function(e,t,n){document.addEventListener("DOMContentLoaded",function(){var e=document.getElementById("product-oneclick-ebanx-button"),t=document.getElementById("ebanx-one-click-tooltip"),n=document.getElementById("ebanx-one-click-close-button"),o=document.getElementsByName("payment[selected_card]"),c=document.getElementById("ebanx-oneclick-pay-button"),r=document.getElementById("product_addtocart_form");o.length&&(o[0].checked=!0),e&&e.addEventListener("click",function(e){e.preventDefault(),t.classList.toggle("is-active")}),n&&n.addEventListener("click",function(e){e.preventDefault(),t.classList.remove("is-active")}),c&&r&&c.addEventListener("click",function(e){r.action="/ebanx/oneclick/pay",productAddToCartForm.submit(c)})})},{}]},{},[1]);
+//# sourceMappingURL=oneclick.js.map
