+++ skin/frontend/base/default/ebanx/js/onestepcheckout.js
@@ -1,19 +1,2 @@
-var ebanxRemoveOSCRequireDocument = function (country) {
-  var doc = document.getElementById('billing:taxvat');
-  var advice = document.getElementById('advice-validate-taxvat-billing:taxvat');
-
-  if (!doc) {
-    return;
-  }
-  if (!doc.classList.contains('validate-taxvat')){
-    doc.classList.add('validate-taxvat');
-  }
-  if (!(country === 'BR' || country === 'CO' || country === 'CL')) {
-    doc.classList.remove('validate-taxvat');
-    doc.classList.remove('validation-failed');
-    if (advice){
-      advice.remove();
-    }
-  }
-
-};
+"use strict";!function r(e,n,t){function o(i,f){if(!n[i]){if(!e[i]){var c="function"==typeof require&&require;if(!f&&c)return c(i,!0);if(u)return u(i,!0);throw new Error("Cannot find module '"+i+"'")}var a=n[i]={exports:{}};e[i][0].call(a.exports,function(r){var n=e[i][1][r];return o(n||r)},a,a.exports,r,e,n,t)}return n[i].exports}for(var u="function"==typeof require&&require,i=0;i<t.length;i++)o(t[i]);return o}({1:[function(r,e,n){},{}]},{},[1]);
+//# sourceMappingURL=onestepcheckout.js.map
