+++ skin/frontend/base/default/ebanx/js/success.js
@@ -1,30 +1,2 @@
-document.addEventListener('DOMContentLoaded', function () {
-  var clipboard = new Clipboard('.ebanx-button--copy');
-
-  clipboard.on('success', function (e) {
-    var text = e.trigger.innerText;
-
-    e.trigger.innerText = 'Copiado!';
-
-    setTimeout(function () {
-      e.trigger.innerText = text;
-    }, 2000);
-  });
-
-  // iFrame Resizer
-  var iframe = document.querySelector('.ebanx-cash-payment iframe');
-
-  if (iframe) {
-    var resizeIframe = function resizeIframe(iframe) {
-      iframe.style.height = iframe.contentWindow.document.body.parentElement.scrollHeight + 'px';
-    }
-
-    window.addEventListener('load', function () {
-      resizeIframe(iframe);
-    });
-
-    iframe.contentWindow.addEventListener('resize', function () {
-      resizeIframe(iframe);
-    });
-  }
-});
+"use strict";!function e(n,t,r){function o(u,c){if(!t[u]){if(!n[u]){var a="function"==typeof require&&require;if(!c&&a)return a(u,!0);if(i)return i(u,!0);throw new Error("Cannot find module '"+u+"'")}var f=t[u]={exports:{}};n[u][0].call(f.exports,function(e){var t=n[u][1][e];return o(t||e)},f,f.exports,e,n,t,r)}return t[u].exports}for(var i="function"==typeof require&&require,u=0;u<r.length;u++)o(r[u]);return o}({1:[function(e,n,t){document.addEventListener("DOMContentLoaded",function(){new Clipboard(".ebanx-button--copy").on("success",function(e){var n=e.trigger.innerText;e.trigger.innerText="Copiado!",setTimeout(function(){e.trigger.innerText=n},2e3)});var e=document.querySelector(".ebanx-cash-payment iframe");if(e){var n=function(e){e.style.height=e.contentWindow.document.body.parentElement.scrollHeight+"px"};window.addEventListener("load",function(){n(e)}),e.contentWindow.addEventListener("resize",function(){n(e)})}})},{}]},{},[1]);
+//# sourceMappingURL=success.js.map
