+++ src/frontend/base/default/ebanx/js/amscheckout-inputs.js
@@ -1,0 +1,3 @@
+var amsCheckoutHandler = function () {
+  return qs('label[for="billing\\:taxvat_number"]');
+};
