+++ src/frontend/base/default/ebanx/js/firecheckout-interest.js
@@ -1,0 +1,3 @@
+var ebanxUpdateFireInterest = function () {
+  checkout.update(checkout.urls.payment_method);
+};
